ebook_system/requirements.txt
anthropic>=0.40.0
fpdf2>=2.7.9
